EVA-4113 Update pipeline to Java21 (#504)

+++ eva-accession-core/src/main/java/uk/ac/ebi/eva/accession/core/batch/policies/IllegalStartSkipPolicy.java
@@ -23,7 +23,7 @@ import uk.ac.ebi.eva.accession.core.exceptions.PositionOutsideOfContigException;
 public class IllegalStartSkipPolicy implements SkipPolicy {
 
     @Override
-    public boolean shouldSkip(Throwable exception, int skipCount) throws SkipLimitExceededException {
+    public boolean shouldSkip(Throwable exception, long skipCount) throws SkipLimitExceededException {
         return exception instanceof PositionOutsideOfContigException;
     }
 }
